In the course of including Fan-Chi Lin's codes for cross-correlating horizontal components I changed the structure of the filtering/whitening process. I removed the scripting parts from the C-driver programs because they are always error prone and very inflexible. After stripping these parts everything that these C-drivers do is reading in sac-files and calling the fortran routines for the filtering. On a later stage this driver level could be skipped by writing a python wrapper for the fortran routines. I put the fortran routines in the 'modules' directory in order to finally get rid of the 'white_outphamp' and 'filter4_f' directories and thus simplifying the build process.
I also put the sac-routines ('read_sac', 'writesac', 'read_sac_header') into
a separate file so that they can be linked to files instead of including them
as source code. several C-codes have to be adapted to use this file instead.

Finally, I removed the MacMakefiles and I introduced the 'makeauto.macro'. This
is supposed to hold parameters that are the same for all Makefiles. 



git-svn-id: http://svn.geo.vuw.ac.nz/svn/noise_correlation/auto/trunk@572 e8761022-d16b-44b9-b4d4-61dd033994a1